+++ CHANGES.md
@@ -39,6 +39,7 @@ Changes:
 - Retrieve lower court information in `nj` #1569
 - Retrieve lower court information in `neb` #1569
 - Retrieve lower court information in `nm` #1569
+- Retrieve lower court information in `or` #1569
 - Update `mdag` scraper, site has changed #1598
 
 Fixes: